+++ smashtheque/models/character.py
@@ -1,0 +1,1 @@
+class Character:
