+++ smashtheque/models/player.py
@@ -1,0 +1,5 @@
+class Player:
+
+  # def __init__(self, attributes: {}):
+  #   self.bot = bot
+
